+++ src/app/Routing/personal-menu/personal-menu.module.ts
@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { PersonalMenuRoutingModule} from './personal-menu-routing.module';
-import { MyGroupsTeacherPageModule } from '../my-groups-teacher-page/my-groups-teacher-page.module';
+import { MyGroupsTeacherPageModule } from '../../my-groups-teacher-page/my-groups-teacher-page.module';
 
 @NgModule({
   declarations: [],
