+++ src/app/Routing/task/task-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { TasksTeacherPageComponent } from '../tasks-teacher-page/tasks-teacher-page.component';
-import { TasksStudentPageComponent } from '../tasks-student-page/tasks-student-page.component';
-import { TaskEditCreatePageComponent } from '../task-edit-create-page/task-edit-create-page.component';
-import { TaskSendingPageComponent } from '../task-sending-page/task-sending-page.component';
-import { TaskFullDescrTeacherComponent } from '../task-full-descr-teacher-page/task-full-descr-teacher.component';
-import { TaskAfterSendingPageComponent } from '../task-after-sending-page/task-after-sending-page.component';
+import { TasksTeacherPageComponent } from '../../tasks-teacher-page/tasks-teacher-page.component';
+import { TasksStudentPageComponent } from '../../tasks-student-page/tasks-student-page.component';
+import { TaskEditCreatePageComponent } from '../../task-edit-create-page/task-edit-create-page.component';
+import { TaskSendingPageComponent } from '../../task-sending-page/task-sending-page.component';
+import { TaskFullDescrTeacherComponent } from '../../task-full-descr-teacher-page/task-full-descr-teacher.component';
+import { TaskAfterSendingPageComponent } from '../../task-after-sending-page/task-after-sending-page.component';
 
 const taskRoutes: Routes = [
   {
